requirements.txt
numpy==2.2.3
openai==0.28.1
python-dotenv==1.0.1
scikit_learn==1.6.1
